requirements-dev.txt
aws-lambda-context==1.1.0
boto3==1.34.11
botocore==1.34.11
jmespath==1.0.1
numpy==1.26.2
pandas==2.1.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.10.0
six==1.16.0
tzdata==2023.4
urllib3==2.0.7
